refactor(upstream): extract pool/connection.rs + pool/lifecycle.rs (lab-kvji.12.4 step 7)
Move UpstreamConnection Debug/Drop/shutdown + UpstreamPool::acquire_peer into
connection.rs; drain_for_swap into lifecycle.rs. pub(super) FLIP (plan §2.1):
shutdown + acquire_peer promoted to pub(super) — called cross-module from
lifecycle/probe/ensure (shutdown) and tools_call/resources_read/prompts_get
(acquire_peer). NO-TOUCH: Drop SIGTERM->SIGKILL->abort ordering and acquire_peer
read-guard scope moved byte-identical (verified). Struct defs stay in pool.rs.
Behavior unchanged. 1423 green incl. process_guard test.

Bead: lab-kvji.12.4
